update category 12/12

+++ WebAppTinhVanCat_aspnetcore/Areas/Blog/Controllers/CategoryController.cs
@@ -7,12 +7,14 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using WebAppTinhVanCat_aspnetcore.Models;
+using WebAppTinhVanCat_aspnetcore.Data;
 
 namespace WebAppTinhVanCat_aspnetcore.Areas.Blog.Controllers
 {
     
     [Area("Blog")]
     [Route("admin/blog/category/[action]/{id?}")]
+    [Authorize(Roles = RoleName.Administrator)]
     public class CategoryController : Controller
     {
         private readonly AppDbContext _context;
@@ -25,9 +27,19 @@ namespace WebAppTinhVanCat_aspnetcore.Areas.Blog.Controllers
         // GET: Blog/Category
         public async Task<IActionResult> Index()
         {
-            var appDbContext = _context.Categories.Include(c => c.ParentCategory);
-            return View(await appDbContext.ToListAsync());
+
+            var qr = (from c in _context.Categories select c)
+                                        .Include(c => c.ParentCategory)
+                                        .Include(c => c.CategoryChildren);
+            var categories = (await qr.ToListAsync()).Where(c => c.ParentCategory == null).ToList();
+
+            var selectList = new SelectList(categories, "Id", "Title");
+
+            return View(categories);
         }
+
+        
+
 
         // GET: Blog/Category/Details/5
         public async Task<IActionResult> Details(int? id)
@@ -48,27 +60,87 @@ namespace WebAppTinhVanCat_aspnetcore.Areas.Blog.Controllers
             return View(category);
         }
 
-        // GET: Blog/Category/Create
-        public IActionResult Create()
+
+        // Source: nguồn 
+        // des: Category đã xử lý
+        //level: cấp danh mục 
+        private void CreateSelectItems(List<Category> source, List<Category> des, int level)
         {
-            ViewData["ParentCategoryId"] = new SelectList(_context.Categories, "Id", "Slug");
+            string prefix = string.Concat(Enumerable.Repeat("--- ", level));
+            foreach (var category in source)
+            {
+                
+                des.Add(new Category() { 
+                    Id = category.Id,
+                    Title = prefix + category.Title
+                });
+                if(category.CategoryChildren?.Count > 0)
+                {
+                    CreateSelectItems(category.CategoryChildren.ToList(), des, level+1);
+                }
+            }
+        }
+
+        // GET: Blog/Category/Create
+        public async Task<IActionResult> Create()
+        {
+            var qr = (from c in _context.Categories select c)
+                                       .Include(c => c.ParentCategory)
+                                       .Include(c => c.CategoryChildren);
+
+            var categories = (await qr.ToListAsync()).Where(c => c.ParentCategory == null).ToList();
+
+            categories.Insert(0, new Category(){
+                    Id = -1,
+                    Title = "Không có danh mục cha"
+            });
+
+            var Items = new List<Category>();
+
+            CreateSelectItems(categories,Items,0 );
+
+            var selectList = new SelectList(Items, "Id", "Title");
+
+            ViewData["ParentCategoryId"] = selectList;
+
+
             return View();
         }
 
         // POST: Blog/Category/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,ParentCategoryId,Title,Content,Slug")] Category category)
         {
             if (ModelState.IsValid)
             {
+                if (category.ParentCategoryId == -1) category.ParentCategoryId = null;
                 _context.Add(category);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["ParentCategoryId"] = new SelectList(_context.Categories, "Id", "Slug", category.ParentCategoryId);
+
+            var qr = (from c in _context.Categories select c)
+                                       .Include(c => c.ParentCategory)
+                                       .Include(c => c.CategoryChildren);
+
+            var categories = (await qr.ToListAsync()).Where(c => c.ParentCategory == null).ToList();
+            categories.Insert(0, new Category()
+            {
+                Id = -1,
+                Title = "Không có danh mục cha"
+
+            });
+
+            var Items = new List<Category>();
+
+            CreateSelectItems(categories, Items, 0);
+
+            var selectList = new SelectList(Items, "Id", "Title");
+
+            ViewData["ParentCategoryId"] = selectList;
+
             return View(category);
         }
 
@@ -85,13 +157,28 @@ namespace WebAppTinhVanCat_aspnetcore.Areas.Blog.Controllers
             {
                 return NotFound();
             }
-            ViewData["ParentCategoryId"] = new SelectList(_context.Categories, "Id", "Slug", category.ParentCategoryId);
+
+            var qr = (from c in _context.Categories select c)
+                                      .Include(c => c.ParentCategory)
+                                      .Include(c => c.CategoryChildren);
+
+            var categories = (await qr.ToListAsync()).Where(c => c.ParentCategory == null).ToList();
+            categories.Insert(0, new Category()
+            {
+                Id = -1,
+                Title = "Không có danh mục cha"
+
+            });
+            var Items = new List<Category>();
+            CreateSelectItems(categories, Items, 0);
+            var selectList = new SelectList(Items, "Id", "Title");
+
+            ViewData["ParentCategoryId"] = selectList;
             return View(category);
         }
 
         // POST: Blog/Category/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+       
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,ParentCategoryId,Title,Content,Slug")] Category category)
@@ -101,10 +188,16 @@ namespace WebAppTinhVanCat_aspnetcore.Areas.Blog.Controllers
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            if (category.ParentCategoryId == category.Id)
+            {
+                ModelState.AddModelError(String.Empty,"Phải chọn danh mục cha khác !");
+            }
+
+            if (ModelState.IsValid && category.ParentCategoryId != category.Id)
             {
                 try
                 {
+                    if (category.ParentCategoryId == -1) category.ParentCategoryId = null;
                     _context.Update(category);
                     await _context.SaveChangesAsync();
                 }
@@ -121,7 +214,27 @@ namespace WebAppTinhVanCat_aspnetcore.Areas.Blog.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["ParentCategoryId"] = new SelectList(_context.Categories, "Id", "Slug", category.ParentCategoryId);
+
+            var qr = (from c in _context.Categories select c)
+                                       .Include(c => c.ParentCategory)
+                                       .Include(c => c.CategoryChildren);
+
+            var categories = (await qr.ToListAsync()).Where(c => c.ParentCategory == null).ToList();
+            categories.Insert(0, new Category()
+            {
+                Id = -1,
+                Title = "Không có danh mục cha"
+
+            });
+
+            var Items = new List<Category>();
+
+            CreateSelectItems(categories, Items, 0);
+
+            var selectList = new SelectList(Items, "Id", "Title");
+
+            ViewData["ParentCategoryId"] = selectList;
+
             return View(category);
         }
 
@@ -149,7 +262,18 @@ namespace WebAppTinhVanCat_aspnetcore.Areas.Blog.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var category = await _context.Categories.FindAsync(id);
+            var category = await _context.Categories
+                            .Include(c => c.CategoryChildren)
+                            .FirstOrDefaultAsync(c => c.Id == id);
+
+            if (category == null) return NotFound();
+
+            foreach (var cCategory in category.CategoryChildren)
+            {
+                cCategory.ParentCategoryId = category.ParentCategoryId;
+            }
+
+
             _context.Categories.Remove(category);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
